Prelim work to help organise controllers
