+++ src/Submissions/Api/Api.csproj
@@ -48,4 +48,8 @@
     </Content>
   </ItemGroup>
 
+  <ItemGroup>
+    <Folder Include="Controllers\Directory\" />
+  </ItemGroup>
+
 </Project>
