+++ src/Submissions/Api/Controllers/Submissions/Domain/RejectController.cs
@@ -1,12 +1,11 @@
 ﻿using System.Threading.Tasks;
 using Biobanks.Submissions.Api.Auth;
-using Biobanks.Submissions.Api.Services.Contracts;
-
+using Biobanks.Submissions.Api.Services.Submissions.Contracts;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Swashbuckle.AspNetCore.Annotations;
 
-namespace Biobanks.Submissions.Api.Controllers.Domain
+namespace Biobanks.Submissions.Api.Controllers.Submissions.Domain
 {
     /// <inheritdoc />
     /// <summary>
